Added: free bath in 10th visit; Automatically modify the appointment value if the pet weights more than 10 kg; automatically identify and modify the appointment based in pet specie (dog or cat); IsPaidInCash column added to appointment table for discount purposes;

+++ TapcatAPI/Controllers/PetController.cs
@@ -88,7 +88,7 @@ public class PetController : ControllerBase
         if (pet == null) return NotFound();
         
         if (updateDto.Name != null) pet.Name = updateDto.Name;
-        if (updateDto.Weight.HasValue) pet.Weight = updateDto.Weight.Value;
+        if (updateDto.Weight.HasValue) pet.Weight = (float)updateDto.Weight.Value;
         if (updateDto.Species != null) pet.Species = updateDto.Species;
         if (updateDto.CustomerId.HasValue) 
         {
