+++ TapcatAPI/Models/Pet.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; } = null!;
-        public double Weight { get; set; }
+        public float Weight { get; set; }
         public string Species { get; set; } = null!;
         public int CustomerId  { get; set; }
